Product week 11

+++ src/main/java/com/piseth/java/school/phoneshop/phoneshop/dto/ModelDTO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class ModelDTO {
 
-    private Integer brandId;
+    private Long brandId;
     private String name;
 }
