+++ src/main/java/com/piseth/java/school/phoneshop/phoneshop/entities/Brand.java
@@ -12,7 +12,8 @@ public class Brand {
     @Id // Error no identifier if we don't use id annotation
     @GeneratedValue(strategy = GenerationType.IDENTITY) // auto generate id for us
     @Column(name = "brand_id")  // customize column
-    private Integer id;     // recommend Long
+    private Long id;     // recommend Long
+
     @Column(name = "brand_name")
     private String name;
